% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EnsemblGeneTable.Hs}
\alias{EnsemblGeneTable.Hs}
\title{Standardized gene list from ENSEMBL (human)}
\format{
A dataframe of ENSEMBL and gene symbols
}
\source{
\url{https://www.ensembl.org/Homo_sapiens/Info/Index}
}
\usage{
EnsemblGeneTable.Hs
}
\description{
A reference of stable gene names for Homo Sapiens
}
\keyword{datasets}
