% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{EnsemblGeneTable.Mm}
\alias{EnsemblGeneTable.Mm}
\title{Standardized gene list from ENSEMBL (mouse)}
\format{
A dataframe of ENSEMBL and gene symbols
}
\source{
\url{https://www.ensembl.org/Mus_musculus/Info/Index}
}
\usage{
EnsemblGeneTable.Mm
}
\description{
A reference of stable gene names for Mus Musculus
}
\keyword{datasets}
