% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FindVariableFeatures.STACAS}
\alias{FindVariableFeatures.STACAS}
\title{FindVariableFeatures.STACAS}
\usage{
\method{FindVariableFeatures}{STACAS}(
  obj,
  nfeat = 1500,
  genesBlockList = "default",
  min.exp = 0.01,
  max.exp = 3
)
}
\arguments{
\item{obj}{A Seurat object containing an expression matrix}

\item{nfeat}{Number of top HVG to be returned}

\item{min.exp}{Minimum average normalized expression for HVG. If lower, the gene will be excluded}

\item{max.exp}{Maximum average normalized expression for HVG. If higher, the gene will be excluded}

\item{genesBlocklist}{Optionally takes a list of vectors of gene names. These genes will be removed from initial HVG set. If set to "default",
STACAS uses its internal list \code{data("genes.blocklist")}.
This is useful to mitigate effect of genes associated with technical artifacts or batch effects
(e.g. mitochondrial, heat-shock response). 
If set to `NULL` no genes will be excluded}
}
\value{
Returns a list of highly variable genes
}
\description{
Select highly variable genes (HVG) from an expression matrix. Genes from a blocklist
(e.g. cell cycling genes, mitochondrial genes) can be excluded from the list of
variable genes, as well as genes with very low or very high average expression
}
