% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{genes.blocklist}
\alias{genes.blocklist}
\title{Genes blocklists for excluding HVGs}
\format{
A list of gene signatures
}
\source{
\url{https://github.com/carmonalab/SignatuR}
}
\usage{
genes.blocklist
}
\description{
A list of gene signatures, including cycling, heat-shock response,
mitochondrial and risobomal genes, interferon response; for mouse and
human. Derived from the SignatuR R package:
\url{https://github.com/carmonalab/SignatuR}
}
\keyword{datasets}
