% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{inconsistent_anchors}
\alias{inconsistent_anchors}
\title{Mark anchor as inconsistent based on cell-type labeling}
\usage{
inconsistent_anchors(anchors, pred.colname)
}
\arguments{
\item{anchors}{anchor object obtained from STACAS function \code{FindAnchors.STACAS} or the Seurat function \code{FindAnchors}.}

\item{pred.colname}{colname of seurat object containing consistent celltype information across all samples/datasets}

\item{scoring}{}
}
\value{
A n integration tree to be passed to the integration function.
}
\description{
Anchors will be consider as inconsistent if they are linking cells classified as different cell populations.
}
\examples{
library(STACAS)
}
