% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{IntegrateData.STACAS}
\alias{IntegrateData.STACAS}
\title{IntegrateData.STACAS}
\usage{
IntegrateData.STACAS(
  anchorset,
  new.assay.name = "integrated",
  features.to.integrate = NULL,
  dims = 1:30,
  k.weight = 100,
  sample.tree = NULL,
  semisupervised = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{anchorset}{A set of anchors calculated using \code{FindAnchors.STACAS}}

\item{new.assay.name}{Assay to store the integrated data}

\item{features.to.integrate}{Which genes to include in the corrected integrated space (def. variable genes)}

\item{dims}{Number of dimensions for local anchor weighting}

\item{k.weight}{Number of neighbors for local anchor weighting. Set \code{k.weight="max"} to disable local weighting}

\item{sample.tree}{Specify the order of integration. See \code{SampleTree.STACAS} to calculate an integration tree.}

\item{semisupervised}{Whether to use cell type label information (if available)}

\item{verbose}{Print progress bar and output}
}
\value{
Returns a \code{Seurat} object with a new integrated Assay, with batch-corrected expression values
}
\description{
Integrate a list of datasets using STACAS anchors. Based on the \code{IntegrateData} function from Seurat.
This function requires that you have calculated a set of integration anchors using \code{FindAnchors.STACAS}.
To perform semi-supervised integration, run \code{FindAnchors.STACAS} with cell type annotations labels.
Integration anchors with inconsistent cell type will be excluded from integration, providing an
integrated space that is partially guided by prior information.
}
