\name{PlotAnchors.STACAS}
\alias{PlotAnchors.STACAS}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Plot Anchor distances with STACAS}
\description{
     The distribution plots generated by \code{PlotAnchors.STACAS} can be useful to inspect the anchor distance distributions between datasets
     and select a threshold for anchor filtering for improved integration. The suggested heuristic threshold is normally rather conservative, you
     may choose to apply a stricter threshold in \code{FilterAnchors.STACAS}
}
\usage{
PlotAnchors.STACAS(ref.anchors = NULL,
    obj.names = NULL,
    dist.thr = NULL,
    dist.pct = 0.8)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ref.anchors}{
    A set of anchors calculated using \code{FindAnchors.STACAS}, containing the pairwise distances between anchors.
  }
  \item{obj.names}{
    An optional vector of names for the datasets being integrated - must have the same length of the number of dataset in \code{ref.anchors}
}
 \item{dist.thr}{
    The distance cutoff for the distribution plots (used only for visualization). If NULL, a suggested threshold is automatically calculated as the \code{dist.pct} percentile of the distance between the two closest datasets.
  }
  \item{dist.pct}{
    If \code{dist.thr} is not specified, calculate the filtering threshold as the \code{dist.pct} percentile of the distance between the two closest datasets (default=0.8)
  }
}
\details{
}
\value{
     Returns a list of grobs for plotting. See the examples below for plotting parameters.
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
   \code{Run.STACAS}

   \code{FindAnchors.STACAS}

   \code{FilterAnchors.STACAS}
}
\examples{
    #PlotAnchors.STACAS requires an anchorset object generated by FindAnchors.STACAS
    library(grid)
    library(gridExtra)

    ref.anchors <- FindAnchors.STACAS(object.list, anchor.features = anchor.features)
    plots <- PlotAnchors.STACAS(ref.anchors)
    g.cols <- 3
    g.rows <- as.integer((length(plots) + 2)/g.cols)
    g <- do.call("arrangeGrob", c(plots, ncol = g.cols, nrow = g.rows))
    ggsave("anchor.dist.mean.pairwise.png", plot = g, width = 6 *
        g.cols, height = 6 * g.rows)
    grid.draw(g)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
