% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SampleTree.STACAS}
\alias{SampleTree.STACAS}
\title{Integration tree generation}
\usage{
SampleTree.STACAS(
  anchorset,
  obj.names = NULL,
  hclust.method = c("average", "single", "complete"),
  usecol = "score",
  method = c("weight.sum", "counts"),
  plot = TRUE
)
}
\arguments{
\item{anchorset}{scored anchors  obtained from \code{FindAnchors.STACAS} and \code{FilterAnchors.STACAS} function}

\item{hclust.method}{clustering method to be used (complete, average, single, ward)}

\item{usecol}{column name to be used to compute sample similarity. Default "score"}

\item{method}{aggregation method to be used among anchors for sample similarity computation. Default: weight.sum}

\item{plot}{logical indicating if dendrogram must be ploted}

\item{dist.hard.thr}{distance threshold used to mimic original STACAS behavior}
}
\value{
A n integration tree to be passed to the integration function.
}
\description{
Build an integration tree by clustering samples in a hierarchical manner. Cumulative scoring among anchor pairs will be used as pairwise similarity criteria of samples.
}
