% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SampleTree.weighted}
\alias{SampleTree.weighted}
\title{Integration tree generation}
\usage{
SampleTree.weighted(
  anchorset,
  hclust.method = "complete",
  usecol = "score",
  dist.hard.thr = NULL,
  method = "cum.sum",
  plot = T
)
}
\arguments{
\item{anchorset}{scored anchors  obtained from \code{scoring_anchors} STACAS function}

\item{hclust.method}{custering method to be used (complete, average, single, ward, etc)}

\item{usecol}{colum name to be used to compute sample similarity. Default "score"}

\item{dist.hard.thr}{distance threshold used to mimic original STACAS behaviour}

\item{method}{agregation method to be used among anchors for sample similarity computation. Default: cum.sum}

\item{plot}{logical indicating if dendrogram must be ploted}
}
\value{
A n integration tree to be passed to the integration function.
}
\description{
Build an integration tree by clustering samples in a hierarchical manner. Cumulative scoring among anchor pairs will be used as pairwise similarity criteria of samples.
}
\examples{
library(STACAS)
}
