% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FilterAnchors.STACAS}
\alias{FilterAnchors.STACAS}
\title{FilterAnchors.STACAS}
\usage{
FilterAnchors.STACAS(
  ref.anchors,
  alpha = 0.5,
  semi_supervised = FALSE,
  dist.pct = 0.5,
  dist.scale.factor = 1,
  beta = 0.5,
  q_boltzmann = 0.8
)
}
\arguments{
\item{ref.anchors}{A set of anchors calculated using \code{FindAnchors.STACAS}, containing the pairwise distances between anchors.}

\item{alpha}{Weight on rPCA distance for rescoring (between 0 and 1).}

\item{semi_supervised}{Use consistency between cell type labels to remove anchors.}

\item{dist.pct}{Center of logistic function, based on quantile value of rPCA distance distribution}

\item{dist.scale.factor}{Scale factor for logistic function (multiplied by SD of rPCA distance distribution)}

\item{beta}{Controls the way in which inconsistent anchors will be penalized.}

\item{q_boltzmann}{Defines a baseline threshold to decide if reject an inconsistent anchor or not}
}
\value{
A new anchor object with reweighted anchors scores, and optionally filtered by consistency of cell type annotation
}
\description{
Re-weighs and filters integration anchors based on pairwise rPCA distance and consistency in cell annotation.
These are calculated using \code{FindAnchors.STACAS}.
}
