% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FindAnchors.STACAS}
\alias{FindAnchors.STACAS}
\title{Find integration anchors using STACAS}
\usage{
FindAnchors.STACAS(
  object.list = NULL,
  assay = NULL,
  reference = NULL,
  anchor.features = 1000,
  genesBlockList = "default",
  dims = 1:10,
  normalization.method = c("LogNormalize", "SCT"),
  k.anchor = 5,
  k.score = 30,
  alpha = 0.8,
  anchor.coverage = 0.5,
  correction.scale = 2,
  cell.labels = NULL,
  label.confidence = 1,
  seed = 123,
  verbose = FALSE
)
}
\arguments{
\item{object.list}{A list of Seurat objects. Anchors will be determined between pairs of objects, 
and can subsequently be used for Seurat dataset integration.}

\item{assay}{A vector containing the assay to use for each Seurat object in object.list.
If not specified, uses the default assay.}

\item{reference}{A vector specifying the (indices of the) objects to be used as a reference during integration.
If NULL (default), all pairwise anchors are found.}

\item{anchor.features}{Can be either: \itemize{
\item{A numeric value. This will call \code{Seurat::SelectIntegrationFeatures} to identify \code{anchor.features}
    genes for anchor finding.}
\item{A pre-calculated vector of integration features to be used for anchor search.}}}

\item{genesBlockList}{If \code{anchor.features} is numeric, \code{genesBlockList} optionally takes a list of vectors of
gene names. These genes will be removed from the integration features. If set to "default",
STACAS uses its internal list \code{data("genes.blocklist")}.
This is useful to mitigate effect of genes associated with technical artifacts or batch effects
(e.g. mitochondrial, heat-shock response).}

\item{dims}{The number of dimensions used for PCA reduction}

\item{normalization.method}{Which normalization method was used to prepare the data - either LogNormalize (default) or SCT}

\item{k.anchor}{The number of neighbors to use for identifying anchors}

\item{k.score}{The number of neighbors to use for scoring anchors}

\item{alpha}{Weight on rPCA distance for rescoring (between 0 and 1).}

\item{anchor.coverage}{Center of logistic function, based on quantile value of rPCA distance distribution}

\item{correction.scale}{Scale factor for logistic function (multiplied by SD of rPCA distance distribution)}

\item{cell.labels}{A metadata column name, storing cell type annotations. These will be taken into account
for semi-supervised alignment (optional). Cells annotated as NA or NULL will not be penalized in semi-supervised
alignment}

\item{label.confidence}{How much you trust the provided cell labels (from 0 to 1).}

\item{seed}{Random seed for probabilistic anchor acceptance}

\item{verbose}{Print all output}
}
\value{
Returns an AnchorSet object, which can be passed to \code{IntegrateData.STACAS}
}
\description{
This function computes anchors between datasets for dataset integration. It is based on the Seurat function
\code{FindIntegrationAnchors}, but is optimized for integration of heterogenous data sets containing only 
partially overlapping cells subsets. It also returns a measure of distance between candidate anchors, 
which can be used at a later stage for anchor filtering using \code{FilterAnchors.STACAS}
}
