% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FindAnchors.STACAS}
\alias{FindAnchors.STACAS}
\title{Find integration anchors using STACAS}
\usage{
FindAnchors.STACAS(
  object.list = NULL,
  assay = NULL,
  reference = NULL,
  min.sample.size = 100,
  max.seed.objects = 10,
  anchor.features = 1000,
  genesBlockList = "default",
  dims = 30,
  k.anchor = 5,
  k.score = 30,
  alpha = 0.8,
  anchor.coverage = 0.5,
  correction.scale = 2,
  cell.labels = NULL,
  label.confidence = 1,
  scale.data = FALSE,
  seed = 123,
  verbose = TRUE
)
}
\arguments{
\item{object.list}{A list of Seurat objects. Anchors will be determined between pairs of objects, 
and can subsequently be used for Seurat dataset integration.}

\item{assay}{A vector containing the assay to use for each Seurat object in object.list.
If not specified, uses the default assay.}

\item{reference}{A vector specifying the object/s to be used as a reference
during integration. If NULL (default), all pairwise anchors are found (no
reference/s). If not NULL, the corresponding objects in \code{object.list}
will be used as references. When using a set of specified references, anchors
are first found between each query and each reference. The references are
then integrated through pairwise integration. Each query is then mapped to
the integrated reference.}

\item{min.sample.size}{Minimum number of cells per sample. Objects with fewer
than this number of cells are not integrated.}

\item{max.seed.objects}{Number of objects to use as seeds to build the integration tree.
Automatically chooses the largest max.seed.objects datasets;
the remaining datasets will be added sequentially to the reference.}

\item{anchor.features}{Can be either: \itemize{
\item{A numeric value. This will call \code{FindVariableFeatures.STACAS} to identify \code{anchor.features}
    that are consistently variable across datasets}
\item{A pre-calculated vector of integration features to be used for anchor search.}}}

\item{genesBlockList}{If \code{anchor.features} is numeric, \code{genesBlockList}
optionally takes a (list of) vectors of gene names. These genes will be
removed from the integration features. If set to "default",
STACAS uses its internal list \code{data("genes.blocklist")}.
This is useful to mitigate effect of genes associated with technical
artifacts or batch effects (e.g. mitochondrial, heat-shock response).}

\item{dims}{The number of dimensions used for PCA reduction}

\item{k.anchor}{The number of neighbors to use for identifying anchors}

\item{k.score}{The number of neighbors to use for scoring anchors}

\item{alpha}{Weight on rPCA distance for rescoring (between 0 and 1).}

\item{anchor.coverage}{Center of logistic function, based on quantile value of rPCA distance distribution}

\item{correction.scale}{Scale factor for logistic function (multiplied by SD of rPCA distance distribution)}

\item{cell.labels}{A metadata column name, storing cell type annotations. These will be taken into account
for semi-supervised alignment (optional). Note that not all cells need to be annotated - please set
unannotated cells as NA or 'unknown' for this column. Cells with NA or 'unknown' cell labels will not be
penalized in semi-supervised alignment.}

\item{label.confidence}{How much you trust the provided cell labels (from 0 to 1).}

\item{scale.data}{Whether to rescale expression data before PCA reduction.}

\item{seed}{Random seed for probabilistic anchor acceptance}

\item{verbose}{Print all output}
}
\value{
Returns an AnchorSet object, which can be passed to \code{IntegrateData.STACAS}
}
\description{
This function computes anchors between datasets for single-cell data integration. It is based on the Seurat function
\code{FindIntegrationAnchors}, but is optimized for integration of heterogenous data sets containing only 
partially overlapping cells subsets. It also computes a measure of distance between candidate anchors (rPCA), 
which is combined with the Seurat's anchor weight by the factor \code{alpha}. Prior knowledge about
cell types can optionally be provided to guide anchor finding.
Give this information in the \code{cell.labels} metadata column. This annotation level, which can be incomplete
(set to NA for cells of unknown type), is used to penalize anchor pairs with inconsistent annotation.
The set of anchors returned by this function can then be passed to \code{IntegrateData.STACAS}
for dataset integration.
}
\examples{
data(sampleObj)
library(Seurat)
obj.list <- SplitObject(sampleObj, split.by="donor")
anchors <- FindAnchors.STACAS(obj.list, min.sample.size=10, k.score=5, dims=3)
}
