% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{FindVariableFeatures.STACAS}
\alias{FindVariableFeatures.STACAS}
\title{Find variable features for STACAS}
\usage{
\method{FindVariableFeatures}{STACAS}(
  obj,
  nfeat = 1500,
  genesBlockList = NULL,
  min.exp = 0.01,
  max.exp = 3
)
}
\arguments{
\item{obj}{A Seurat object containing an expression matrix}

\item{nfeat}{Number of top HVG to be returned}

\item{min.exp}{Minimum average normalized expression variable for HVG. If lower, the gene will be excluded}

\item{max.exp}{Maximum average normalized expression variable for HVG. If higher, the gene will be excluded}

\item{genesBlocklist}{A list of genes to be excluded from the resulting HVG}
}
\value{
Returns a list of highly variable genes
}
\description{
Select highly variable genes (HVG) from an expression matrix. Genes from a blocklist
(e.g. cell cycling genes, mitochondrial genes) can be excluded from the list of
variable genes, as well as genes with very low or very high average expression.
}
