% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{IntegrateData.STACAS}
\alias{IntegrateData.STACAS}
\title{IntegrateData.STACAS}
\usage{
IntegrateData.STACAS(
  anchorset,
  new.assay.name = "integrated",
  normalization.method = c("LogNormalize", "SCT"),
  features.to.integrate = NULL,
  dims = 1:30,
  k.weight = 100,
  sample.tree = NULL,
  semisupervised = TRUE,
  preserve.order = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{anchorset}{A set of anchors calculated using \code{FindAnchors.STACAS} and filtered using \code{FilterAnchors.STACAS}}

\item{new.assay.name}{Assay to store the integrated data}

\item{normalization.method}{How was the data normalized?}

\item{features.to.integrate}{Which genes to include in the corrected integrated space (def. variable genes)}

\item{dims}{Number of dimensions for local anchor weighting}

\item{k.weight}{Number of neighbors for local anchor weighting. Set \code{k.weight="max"} to disable local weighting}

\item{sample.tree}{Specify the order of integration. See \code{SampleTree.STACAS} to calculate an integration tree.}

\item{semisupervised}{Whether to use cell type label information (if available)}

\item{preserve.order}{Do not reorder objects based on size for each pairwise integration.}

\item{verbose}{Print progress bar and output}
}
\value{
Returns a \code{Seurat} object with a new integrated Assay. If normalization.method = "LogNormalize", the integrated data is returned to the data slot and can be treated as log-normalized, corrected data. If normalization.method = "SCT", the integrated data is returned to the scale.data slot and can be treated as centered, corrected Pearson residuals.
}
\description{
Integrate a list of datasets using STACAS anchors
}
