% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Run.STACAS}
\alias{Run.STACAS}
\title{Run the STACAS integration pipeline}
\usage{
Run.STACAS(
  object.list = NULL,
  assay = NULL,
  anchor.features = 1000,
  genesBlockList = "default",
  dims = 1:30,
  k.anchor = 5,
  k.score = 30,
  k.weight = 100,
  alpha = 0.8,
  anchor.coverage = 0.5,
  correction.scale = 2,
  cell.labels = NULL,
  label.confidence = 1,
  hclust.method = NULL,
  seed = 123,
  verbose = FALSE
)
}
\arguments{
\item{object.list}{A list of Seurat objects. Anchors will be determined between pairs of objects, 
and can subsequently be used for Seurat dataset integration.}

\item{assay}{A vector containing the assay to use for each Seurat object in object.list.
If not specified, uses the default assay.}

\item{anchor.features}{Can be either: \itemize{
\item{A numeric value. This will call \code{Seurat::SelectIntegrationFeatures} to identify \code{anchor.features}
    genes for anchor finding.}
\item{A pre-calculated vector of integration features to be used for anchor search.}}}

\item{genesBlockList}{If \code{anchor.features} is numeric, \code{genesBlockList} optionally takes a list of vectors of
gene names. These genes will be removed from the integration features. If set to "default",
STACAS uses its internal list \code{data("genes.blocklist")}.
This is useful to mitigate effect of genes associated with technical artifacts or batch effects
(e.g. mitochondrial, heat-shock response).}

\item{dims}{The number of dimensions used for PCA reduction}

\item{k.anchor}{The number of neighbors to use for identifying anchors}

\item{k.score}{The number of neighbors to use for scoring anchors}

\item{k.weight}{Number of neighbors for local anchor weighting. Set \code{k.weight="max"} to disable local weighting}

\item{alpha}{Weight on rPCA distance for rescoring (between 0 and 1).}

\item{anchor.coverage}{Center of logistic function, based on quantile value of rPCA distance distribution}

\item{correction.scale}{Scale factor for logistic function (multiplied by SD of rPCA distance distribution)}

\item{cell.labels}{A metadata column name, storing cell type annotations. These will be taken into account
for semi-supervised alignment (optional). Cells annotated as NA or NULL will not be penalized in semi-supervised
alignment}

\item{label.confidence}{How much you trust the provided cell labels (from 0 to 1).}

\item{hclust.method}{Clustering method for integration tree (single, complete, average, ward)}

\item{seed}{Random seed for probabilistic anchor acceptance}

\item{verbose}{Print all output}
}
\value{
Returns a \code{Seurat} object with a new integrated Assay. Also, centered, scaled variable features data are returned in the scale.data slot, and the pca of these batch-corrected scale data in the pca `reduction` slot
}
\description{
This function is a wrapper for running the several steps required to integrate single-cell
datasets using STACAS: 1) Finding integration anchors; 2) Calculating the sample tree for
the order of dataset integration; 3) Dataset batch effect correction and integration
}
