% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SampleTree.STACAS}
\alias{SampleTree.STACAS}
\title{Integration tree generation}
\usage{
SampleTree.STACAS(
  anchorset,
  obj.names = NULL,
  hclust.method = c("average", "single", "complete"),
  usecol = "score",
  method = c("weight.sum", "counts"),
  semisupervised = TRUE,
  plot = TRUE
)
}
\arguments{
\item{anchorset}{Scored anchorsobtained from \code{FindAnchors.STACAS} and \code{FilterAnchors.STACAS} function}

\item{hclust.method}{Clustering method to be used (complete, average, single, ward)}

\item{usecol}{Column name to be used to compute sample similarity. Default "score"}

\item{method}{Aggregation method to be used among anchors for sample similarity computation. Default: weight.sum}

\item{semisupervised}{Whether to use cell type label information (if available)}

\item{plot}{Logical indicating if dendrogram must be plotted}

\item{dist.hard.thr}{Distance threshold used to mimic original STACAS behavior}
}
\value{
An integration tree to be passed to the integration function.
}
\description{
Build an integration tree by clustering samples in a hierarchical manner. Cumulative scoring among anchor pairs will be used as pairwise similarity criteria of samples.
}
