% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{StandardizeGeneSymbols}
\alias{StandardizeGeneSymbols}
\title{Standardize gene symbols}
\usage{
StandardizeGeneSymbols(obj, EnsemblGeneTable = NULL, EnsemblGeneFile = NULL)
}
\arguments{
\item{obj}{A Seurat object}

\item{EnsemblGeneTable}{A data frame of gene name mappings. This should have
the format of \href{https://www.ensembl.org/info/data/biomart/index.html}{Ensembl BioMart tables}
with fields "Gene name", "Gene Synonym" and "Gene stable ID" (and optionally
"NCBI gene (formerly Entrezgene) ID"). See also
the default conversion table in STACAS with \code{data(EnsemblGenes105_Mmu_GRCm39)}}

\item{EnsemblGeneFile}{If \code{EnsemblGeneTable==NULL}, read a gene mapping
table from this file}
}
\value{
Returns a Seurat object with standard gene names. Genes not found in
    the standard list are removed. Synonyms are accepted when
    the conversion is not ambiguous.
}
\description{
Converts gene names of a Seurat single-cell object to a dictionary of
standard symbols. This function is useful prior to integration of datasets
from different studies, where gene names may be inconsistent.
}
